requirements.txt
# Web Framework & Core
Flask==3.1.2
gunicorn
Werkzeug==3.1.5
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0
MarkupSafe==3.0.3
 
# Database & Data
psycopg2-binary==2.9.11
pandas==3.0.0
numpy==2.4.1
python-dotenv==1.2.1
folium==0.20.0
seaborn==0.13.2
 
# Utility
python-dateutil==2.9.0.post0
typing_extensions==4.15.0
tzdata==2025.3
packaging==26.0
fonttools==4.61.1
charset-normalizer==3.4.4
decorator==5.2.1